fix: ignore string-based plugins in md
